Website Load check
I have put this file in my hosting and tried to use in my website mail-ship.com. As the website is empty, I set the 0.008 sec as a limit to check if the warning works. I added the webhook of the slack channel and succesfully received the warning in Slack from my side.